+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add customer credit information on totalizers
+
+### Added
+
 - Node and graphql setup for saved carts
 
 ### Changed